+++ src/test/java/com/ververica/flink/table/gateway/operation/DescribeOperationTest.java
@@ -19,24 +19,16 @@
 package com.ververica.flink.table.gateway.operation;
 
 import com.ververica.flink.table.gateway.config.Environment;
-import com.ververica.flink.table.gateway.rest.result.ColumnInfo;
-import com.ververica.flink.table.gateway.rest.result.ConstantNames;
-import com.ververica.flink.table.gateway.rest.result.ResultKind;
 import com.ververica.flink.table.gateway.rest.result.ResultSet;
-import com.ververica.flink.table.gateway.rest.result.TableSchemaUtil;
 import com.ververica.flink.table.gateway.utils.EnvironmentFileUtil;
 
 import org.apache.flink.table.api.DataTypes;
 import org.apache.flink.table.api.TableSchema;
-import org.apache.flink.table.types.logical.VarCharType;
-import org.apache.flink.types.Row;
 
 import org.junit.Test;
 
 import java.util.HashMap;
 import java.util.Map;
-
-import static org.junit.Assert.assertEquals;
 
 /**
  * Tests for {@link DescribeOperation}.
@@ -63,19 +55,6 @@ public class DescribeOperationTest extends OperationTestBase {
 			.field("IntegerField1", DataTypes.INT())
 			.field("StringField1", DataTypes.STRING())
 			.build();
-
-		compareResult(tableSchema, resultSet);
-	}
-
-	public static void compareResult(TableSchema tableSchema, ResultSet resultSet) throws Exception {
-		String schemaJson = TableSchemaUtil.writeTableSchemaToJson(tableSchema);
-
-		ResultSet expected = ResultSet.builder()
-			.resultKind(ResultKind.SUCCESS_WITH_CONTENT)
-			.columns(ColumnInfo.create(ConstantNames.SCHEMA, new VarCharType(false, schemaJson.length())))
-			.data(Row.of(schemaJson))
-			.build();
-
-		assertEquals(expected, resultSet);
+		OperationTestUtils.compareDescribeResult(tableSchema, resultSet);
 	}
 }
